+++ src/answers.ts
@@ -1,38 +1,39 @@
-const logger = require('./logger');
+import type { Req, Res } from './server';
+import * as config from './config';
+import logger from './logger';
 
-async function get(req, res, next) {
+export async function get(req: Req, res: Res, next) {
   let status = 200;
   let html;
   let type = 'text/plain';
   switch (req.url) {
     case '/security.txt':
     case '/.well-known/security.txt':
-      // @ts-ignore
-      html = global.config['security.txt'];
+      html = config.get().answers['security.txt'];
       break;
     case '/robots.txt':
     case '/.well-known/robots.txt':
-      // @ts-ignore
-      html = global.config['robots.txt'];
+      html = config.get().answers['robots.txt'];
       break;
     case '/humans.txt':
     case '/.well-known/humans.txt':
-      // @ts-ignore
-      html = global.config['humans.txt'];
-      status = 418;
+      html = config.get().answers['humans.txt'];
       break;
     case '/sitemap.xml':
     case '/.well-known/sitemap.xml':
-      const host = req.headers[':authority'] || req.headers.host;
-      // @ts-ignore
-      html = (global.config['sitemap.xml'] || '').replace('URL', host);
+      const host: string = req.headers[':authority']?.toString() || req.headers.host?.toString() || '';
+      html = ((config.get().answers['sitemap.xml'] || '') as string).replace('URL', host);
       type = 'text/xml';
       status = 451;
       break;
     case '/.git/HEAD':
-      // @ts-ignore
-      html = global.config['git.head'];
+      html = config.get().answers['git.head'];
       status = 403;
+      break;
+    case '/ver':
+    case '/version':
+      html = JSON.stringify(config.get().answers['version'], null, 2);
+      type = 'application/json';
       break;
     default:
       next();
@@ -43,10 +44,8 @@ async function get(req, res, next) {
     return;
   }
   res.writeHead(status, { 'Content-Type': `'${type}'`, 'Cache-Control': 'no-cache', 'X-Content-Type-Options': 'nosniff' });
-  logger(req, res);
+  logger(req, req);
   res.end(html, 'utf-8');
 }
-
-exports.get = get;
 
 // https://en.wikipedia.org/wiki/List_of_HTTP_status_codes