+++ src/proxy.ts
@@ -1,9 +1,13 @@
-/* based on https://github.com/nxtedition/node-http2-proxy by https://github.com/ronag */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable no-use-before-define */
 
-const net = require('net');
-const http = require('http');
-const https = require('https');
+/* based on https://github.com/nxtedition/node-http2-proxy by https://github.com/ronag */
+
+import * as net from 'net';
+import * as http from 'http';
+import * as https from 'https';
+import type { TlsOptions, TLSSocket } from 'tls';
+import type { Req, Res, Headers } from './server';
 
 const tlsOptions = [
   'ca',
@@ -37,10 +41,10 @@ const kOnRes = Symbol('onRes');
 
 async function request({ req, socket, res = socket, head, proxyName }, onReq, onRes) {
   if (req.aborted) return null;
-  const headers = getRequestHeaders(req, proxyName);
-  if (head !== undefined) {
-    if (req.method !== 'GET') throw new HttpError('only GET request allowed', null, 405);
-    if (req.headers['upgrade'] !== 'websocket') throw new HttpError('missing upgrade header', null, 400);
+  const headers: Headers = getRequestHeaders(req, proxyName);
+  if (head) {
+    if (req.method !== 'GET') throw new HttpError('only GET request allowed', undefined, 405);
+    if (req.headers['upgrade'] !== 'websocket') throw new HttpError('missing upgrade header', undefined, 400);
     if (head && head.length) res.unshift(head);
     setupSocket(res);
     headers['connection'] = 'upgrade';
@@ -48,8 +52,13 @@ async function request({ req, socket, res = socket, head, proxyName }, onReq, on
   }
   const proxyReq = await onReq({ method: req.method, path: req.originalUrl || req.url, headers });
   if (req.aborted) {
-    if (proxyReq.abort) proxyReq.abort();
-    else if (proxyReq.destroy) proxyReq.destroy();
+    if (proxyReq?.abort) proxyReq.abort();
+    else if (proxyReq?.destroy) proxyReq.destroy();
+    return null;
+  }
+  if (req.headers['unknown-protocol']) {
+    res.statusCode = 421;
+    res.end('uknown target');
     return null;
   }
   let callback;
@@ -62,6 +71,7 @@ async function request({ req, socket, res = socket, head, proxyName }, onReq, on
   res[kRes] = res;
   res[kProxySocket] = null;
   res[kProxyRes] = null;
+  // @ts-ignore callback is assigned in promise
   res[kProxyCallback] = callback;
   proxyReq[kReq] = req;
   proxyReq[kRes] = res;
@@ -84,17 +94,17 @@ async function request({ req, socket, res = socket, head, proxyName }, onReq, on
   return promise;
 }
 
-function onSocket(socket) {
+function onSocket(this: any, socket: TLSSocket) {
   if (!socket.connecting) onProxyConnect.call(this);
   else socket.once('connect', onProxyConnect.bind(this));
 }
 
-function deferToConnect() {
+function deferToConnect(this: any) { // eslint-disable-line no-unused-vars
   if (this.socket) onSocket.call(this, this.socket);
   else this.once('socket', onSocket);
 }
 
-function onComplete(err) {
+function onComplete(this: any, err) {
   const res = this[kRes];
   const req = res[kReq];
   if (!res[kProxyCallback]) return;
@@ -130,48 +140,46 @@ function onComplete(err) {
   callback(err);
 }
 
-function onProxyConnect() {
+function onProxyConnect(this: any) { // eslint-disable-line no-unused-vars
   this[kConnected] = true;
   if (this['method'] === 'GET' || this['method'] === 'HEAD' || this['method'] === 'OPTIONS') {
     this[kReq].resume();
-    // @ts-ignore
     this.end();
   } else {
     this[kReq]
       .on('data', onReqData)
       .on('end', onReqEnd);
     this
-      // @ts-ignore
       .on('drain', onProxyReqDrain);
   }
 }
 
-function onReqEnd() {
+function onReqEnd(this: any) { // eslint-disable-line no-unused-vars
   this[kProxyReq].end();
 }
 
-function onReqData(buf) {
+function onReqData(this: any, buf) {
   if (!this[kProxyReq].write(buf)) this.pause();
 }
 
-function onProxyReqDrain() {
+function onProxyReqDrain(this: any) { // eslint-disable-line no-unused-vars
   this[kReq].resume();
 }
 
-function onProxyReqError(err) {
+function onProxyReqError(this: any, err) {
   err.statusCode = this[kConnected] ? 502 : 503;
   onComplete.call(this, err);
 }
 
-function onProxyReqTimeout() {
+function onProxyReqTimeout(this: any) { // eslint-disable-line no-unused-vars
   onComplete.call(this, new HttpError('proxy timeout', 'ETIMEDOUT', 504));
 }
 
-async function onProxyReqResponse(proxyRes) {
+async function onProxyReqResponse(this: any, proxyRes) {
   const res = this[kRes];
   res[kProxyRes] = proxyRes;
   proxyRes[kRes] = res;
-  const headers = setupHeaders(proxyRes.headers);
+  const headers: Headers = setupHeaders(proxyRes.headers);
   proxyRes.on('aborted', onProxyResAborted).on('error', onProxyResError);
   if (this[kOnRes]) {
     try {
@@ -191,7 +199,7 @@ async function onProxyReqResponse(proxyRes) {
   }
 }
 
-function onProxyReqUpgrade(proxyRes, proxySocket, proxyHead) {
+function onProxyReqUpgrade(this: any, proxyRes, proxySocket, proxyHead) {
   const res = this[kRes];
   res[kProxySocket] = proxySocket;
   proxySocket[kRes] = res;
@@ -205,35 +213,33 @@ function onProxyReqUpgrade(proxyRes, proxySocket, proxyHead) {
     .pipe(proxySocket);
 }
 
-function onProxyResError(err) {
+function onProxyResError(this: any, err) {
   err.statusCode = 502;
   onComplete.call(this, err);
 }
 
-function onProxyResAborted() {
+function onProxyResAborted(this: any) { // eslint-disable-line no-unused-vars
   onComplete.call(this, new HttpError('proxy aborted', 'ECONNRESET', 502));
 }
 
-function onProxyResEnd() {
+function onProxyResEnd(this: any) { // eslint-disable-line no-unused-vars
   if (this.trailers) this[kRes].addTrailers(this.trailers);
 }
 
-function createHttpHeader(line, headers) {
+function createHttpHeader(line: string, headers: Headers) {
   let head = line;
   for (const [key, value] of Object.entries(headers)) {
     if (!Array.isArray(value)) {
       head += `\r\n${key}: ${value}`;
     } else {
-      for (let i = 0; i < value.length; i++) {
-        head += `\r\n${key}: ${value[i]}`;
-      }
+      for (let i = 0; i < value.length; i++) head += `\r\n${key}: ${value[i]}`;
     }
   }
   head += '\r\n\r\n';
   return Buffer.from(head, 'ascii');
 }
 
-function getRequestHeaders(req, proxyName) {
+function getRequestHeaders(req: { headers: Headers; httpVersion: any; socket: { localAddress: any; localPort: any; remoteAddress: any; remotePort: any; encrypted: any; }; }, proxyName: any) {
   const headers = {};
   for (const [key, value] of Object.entries(req.headers)) {
     if (key.charAt(0) !== ':' && key !== 'host') headers[key] = value;
@@ -242,7 +248,7 @@ function getRequestHeaders(req, proxyName) {
   if (proxyName) {
     if (headers['via']) {
       for (const name of headers['via'].split(',')) {
-        if (name.endsWith(proxyName)) throw new HttpError('loop detected', null, 508);
+        if (name.endsWith(proxyName)) throw new HttpError('loop detected', undefined, 508);
       }
       headers['via'] += ',';
     } else {
@@ -251,7 +257,7 @@ function getRequestHeaders(req, proxyName) {
     headers['via'] += `${req.httpVersion} ${proxyName}`;
   }
 
-  function printIp(address, port) {
+  function printIp(address: string, port?: number) {
     const isIPv6 = net.isIPv6(address);
     let str = `${address}`;
     if (isIPv6) str = `[${str}]`;
@@ -264,23 +270,22 @@ function getRequestHeaders(req, proxyName) {
     `by=${printIp(req.socket.localAddress, req.socket.localPort)}`,
     `for=${printIp(req.socket.remoteAddress, req.socket.remotePort)}`,
     `proto=${req.socket.encrypted ? 'https' : 'http'}`,
-    `host=${printIp(req.headers[':authority'] || req.headers['host'] || '')}`,
+    `host=${printIp(req.headers[':authority'] as string || req.headers['host'] as string || '')}`,
   ].join(';');
   if (headers['forwarded']) headers['forwarded'] += `, ${forwarded}`;
   else headers['forwarded'] = `${forwarded}`;
   return setupHeaders(headers);
 }
 
-function setupSocket(socket) {
+function setupSocket(socket: TLSSocket) {
   socket.setTimeout(0);
   socket.setNoDelay(true);
   socket.setKeepAlive(true, 0);
 }
 
-function setupHeaders(headers) {
-  const connection = headers['connection'];
-  if (connection && connection !== 'connection' && connection !== 'keep-alive') {
-    for (const name of connection.toLowerCase().split(',')) delete headers[name.trim()];
+function setupHeaders(headers: Headers) {
+  if (headers['connection'] && headers['connection'] !== 'connection' && headers['connection'] !== 'keep-alive') {
+    for (const name of headers['connection'].toLowerCase().split(',')) delete headers[name.trim()];
   }
   delete headers['connection'];
   delete headers['proxy-connection'];
@@ -296,15 +301,26 @@ function setupHeaders(headers) {
 }
 
 class HttpError extends Error {
-  constructor(msg, code, statusCode) {
+  code: string | undefined;
+  statusCode: number;
+
+  constructor(msg: string | undefined, code: string | undefined, statusCode: number) {
     super(msg);
     this.code = code;
     this.statusCode = statusCode || 500;
   }
 }
 
-function process(proxy) {
-  proxy.ws = function ws(req, socket, head, options, callback) {
+type Proxy = {
+  ({ req, socket, res, head, proxyName }: { req: Req; socket: TlsOptions; res?: Res; head: Headers; proxyName: string; }, onReq, onRes): Promise<unknown>; // eslint-disable-line no-unused-vars
+  (arg0, arg1, arg2); // eslint-disable-line no-unused-vars
+  (arg0, arg1, arg2); // eslint-disable-line no-unused-vars
+  ws?;
+  web?;
+}
+
+export function process(proxy: Proxy) {
+  proxy.ws = function ws(req: Req, socket: TLSSocket, head: Headers, options: any, callback) {
     const promise = compat({ req, socket, head }, options);
     if (!callback) return promise;
     return promise
@@ -312,55 +328,58 @@ function process(proxy) {
       .catch((err) => callback(err, req, socket, head));
   };
 
-  proxy.web = function web(req, res, options, callback) {
+  proxy.web = function web(req: Req, res: Res, options: any, callback) {
     const promise = compat({ req, res }, options);
     if (!callback) return promise;
     return promise
       .then(() => callback(null, req, res))
       .catch((err) => callback(err, req, res));
   };
-  async function compat(ctx, options) {
+
+  type Options = {
+    hostname: string,
+    port: number,
+    path: string,
+    socketPath: string,
+    protocol: string,
+    timeout: number,
+    proxyTimeout: number,
+    proxyName: string,
+    onReq,
+    onRes,
+  }
+
+  async function compat(ctx: { req: Req; socket?: TLSSocket; head?: Headers; res?: Res; }, options: Options) {
     const { req, res } = ctx;
-    const {
-      hostname,
-      port,
-      path,
-      socketPath,
-      protocol,
-      timeout,
-      proxyTimeout,
-      proxyName,
-      onReq,
-      onRes,
-    } = options;
-    if (timeout != null) req.setTimeout(timeout);
+    const { hostname, port, path, socketPath, protocol, timeout, proxyTimeout, proxyName, onReq, onRes } = options;
+    if (timeout) req.setTimeout(timeout);
     await proxy(
       { ...ctx, proxyName },
-      async (ureq) => {
+      async (ureq: https.RequestOptions) => {
         for (const key of tlsOptions) {
           if (Reflect.has(options, key)) {
             const value = Reflect.get(options, key);
             Reflect.set(ureq, key, value);
           }
         }
-        if (hostname !== undefined) ureq.hostname = hostname;
-        if (port !== undefined) ureq.port = port;
-        if (path !== undefined) ureq.path = path;
-        if (proxyTimeout !== undefined) ureq.timeout = proxyTimeout;
-        if (socketPath !== undefined) ureq.socketPath = socketPath;
-        let ret;
+        if (hostname) ureq.hostname = hostname;
+        if (port) ureq.port = port;
+        if (path) ureq.path = path;
+        if (proxyTimeout) ureq.timeout = proxyTimeout;
+        if (socketPath) ureq.socketPath = socketPath;
+        let ret: http.ClientRequest | undefined;
         if (onReq) {
           if (onReq.length <= 2) {
             ret = await onReq(req, ureq);
           } else {
             ret = await new Promise((resolve, reject) => {
-              const promiseOrReq = onReq(req, ureq, (err, val) => (err ? reject(err) : resolve(val)));
+              const promiseOrReq = onReq(req, ureq, (err, val: http.ClientRequest | Promise<http.ClientRequest>) => (err ? reject(err) : resolve(val)));
               if (promiseOrReq) {
                 if (promiseOrReq.then) promiseOrReq.then(resolve).catch(reject);
                 else if (promiseOrReq.abort) resolve(promiseOrReq);
                 else throw new Error('onReq must return a promise or a request object');
               } else {
-                resolve(true);
+                reject();
               }
             });
           }
@@ -370,11 +389,12 @@ function process(proxy) {
           if (protocol == null || /^(http|ws):?$/.test(protocol)) agent = http;
           else if (/^(http|ws)s:?$/.test(protocol)) agent = https;
           else throw new Error('invalid protocol');
-          ret = agent.request(ureq);
+          if (ureq.hostname && ureq.port) ret = agent.request(ureq);
+          else req.headers['unknown-protocol'] = 'true';
         }
         return ret;
       },
-      onRes ? async (proxyRes, headers) => {
+      onRes ? async (proxyRes: { headers: Headers }, headers: Headers) => {
         proxyRes.headers = headers;
         if (onRes.length <= 3) return onRes(req, res, proxyRes);
         return new Promise((resolve, reject) => {
@@ -389,4 +409,4 @@ function process(proxy) {
   return proxy;
 }
 
-module.exports = process(request);
+export default process(request);