+++ CHANGELOG.md
@@ -1,156 +1,219 @@
-# PiGallery Change Log
+# @vladmandic/piproxy  
 
-### **HEAD -> master, origin/master, origin/HEAD** 2022/08/11 mandic00@live.com
+  Version: **2.0.0**  
+  Description: **Reverse HTTP/HTTPS/HTTP2 Proxy in NodeJS with SSL Certificate Management, Compression, Security Enforcement and Rate Limiting**  
+  
+  Author: **Vladimir Mandic <mandic00@live.com>**  
+  License: **MIT**  
+  Repository: **<https://github.com/vladmandic/piproxy>**  
+  
+## Changelog
+  
+### **1.1.24** 2022/08/11 mandic00@live.com
+
+- fix loading of service workers
 
 ### **1.1.23** 2022/05/25 mandic00@live.com
 
+
 ### **1.1.22** 2022/03/20 mandic00@live.com
+
 
 ### **1.1.21** 2021/12/14 mandic00@live.com
 
+
 ### **1.1.20** 2021/10/02 mandic00@live.com
+
 
 ### **1.1.19** 2021/07/29 mandic00@live.com
 
-### **tag: 1.1.8** 2021/04/22 mandic00@live.com
+
+### **release: 1.1.8** 2021/04/22 mandic00@live.com
+
 
 ### **1.1.18** 2021/04/21 mandic00@live.com
+
 - fix frame-ancesor for iframes
 - reimplement header handling
 
 ### **1.1.17** 2021/04/07 mandic00@live.com
 
+
 ### **1.1.16** 2021/03/18 mandic00@live.com
+
 
 ### **1.1.15** 2021/03/13 mandic00@live.com
 
+
 ### **1.1.14** 2021/02/20 mandic00@live.com
 
-### **tag: v1.1.13** 2021/02/19 mandic00@live.com
 
 ### **1.1.13** 2021/02/19 mandic00@live.com
 
+
 ### **1.1.11** 2021/01/29 mandic00@live.com
+
 
 ### **1.1.10** 2021/01/22 mandic00@live.com
 
+
 ### **1.1.9** 2021/01/22 mandic00@live.com
+
 - strict linting
 
 ### **1.1.8** 2021/01/03 mandic00@live.com
+
 - merge branch 'master' of https://github.com/vladmandic/piproxy
 - merge pull request #3 from vladmandic/dependabot/npm_and_yarn/ini-1.3.8
 - bump ini from 1.3.5 to 1.3.8
 
 ### **1.1.7** 2020/12/12 mandic00@live.com
+
 - port update
 - package update
 
 ### **1.1.6** 2020/11/25 mandic00@live.com
 
+
 ### **1.1.5** 2020/11/25 mandic00@live.com
 
-### **tag: v2** 2020/10/29 mandic00@live.com
+
+### **2** 2020/10/29 mandic00@live.com
+
 
 ### **1.1.4** 2020/10/28 mandic00@live.com
 
+
 ### **1.1.3** 2020/09/26 mandic00@live.com
+
 
 ### **1.1.2** 2020/09/16 cyan00@gmail.com
 
+
 ### **1.1.1** 2020/09/16 cyan00@gmail.com
 
+
 ### **1.0.24** 2020/09/16 cyan00@gmail.com
+
 - improved error handler
 
 ### **1.0.23** 2020/09/11 cyan00@gmail.com
 
+
 ### **1.0.22** 2020/09/11 cyan00@gmail.com
+
 - autorestart on ssl cert update
 
 ### **1.0.21** 2020/09/05 cyan00@gmail.com
+
 - fix firefox compression compatibility issue
 
 ### **1.0.20** 2020/09/04 cyan00@gmail.com
+
 - add gzip as alternative
 
 ### **1.0.19** 2020/09/03 cyan00@gmail.com
 
+
 ### **1.0.18** 2020/08/28 cyan00@gmail.com
+
 - excluded js map files
 - monitoring and stats improvements
 
 ### **1.0.17** 2020/08/22 cyan00@gmail.com
+
 - reduced dependencies
 - tweaks
 
 ### **1.0.16** 2020/08/22 cyan00@gmail.com
+
 - econding fix
 - compression update
 
 ### **1.0.15** 2020/08/19 cyan00@gmail.com
+
 - added geoip db to git lfs
 - added
 
 ### **1.0.14** 2020/08/19 cyan00@gmail.com
+
 - automatic server monitoring
 - added predefined responses
 
 ### **1.0.13** 2020/08/19 cyan00@gmail.com
+
 - improved logging
 
 ### **1.0.12** 2020/08/16 cyan00@gmail.com
 
+
 ### **1.0.11** 2020/08/14 cyan00@gmail.com
+
 - url query
 - enhanced stats
 - improved db logging
 
 ### **1.0.10** 2020/08/13 cyan00@gmail.com
+
 - initial work on stats page
 
 ### **1.0.9** 2020/08/11 cyan00@gmail.com
+
 - fix statuscode
 - ip resolving fix
 
 ### **1.0.8** 2020/08/10 cyan00@gmail.com
+
 - implemented brotli compression
 
 ### **1.0.7** 2020/08/10 cyan00@gmail.com
+
 - priviledge drop
 
 ### **1.0.6** 2020/08/08 cyan00@gmail.com
+
 - simplified dependencies
 
 ### **1.0.5** 2020/08/06 cyan00@gmail.com
+
 - improved configuration options
 
 ### **1.0.4** 2020/08/06 cyan00@gmail.com
 
+
 ### **1.0.3** 2020/08/02 cyan00@gmail.com
+
 - custom error handler
 
 ### **1.0.2** 2020/08/02 cyan00@gmail.com
+
 - added helmet protection and rate limiter
 - rewrite using native methods
 
 ### **1.0.1** 2020/08/01 cyan00@gmail.com
+
 - added secure server redirect
 
 ### **0.0.6** 2020/07/29 cyan00@gmail.com
 
-### **tag: 1.0.0** 2020/07/19 cyan00@gmail.com
+
+### **release: 1.0.0** 2020/07/19 cyan00@gmail.com
+
 
 ### **0.0.5** 2020/07/19 cyan00@gmail.com
+
 - http2 logging
 
 ### **0.0.4** 2020/07/02 cyan00@gmail.com
 
+
 ### **0.0.3** 2020/06/28 cyan00@gmail.com
+
 - noip refresh bugfix
 - deleted logs
 - added logger with geo location and agent analysis
 
 ### **0.0.2** 2020/06/27 cyan00@gmail.com
+
 - readme and systemd file update
 - initial commit